requirements.txt
selenium==2.53.5
xvfbwrapper==0.2.8
pyaml
